src/main/java/racingcar/model/Car.java
package racingcar.model;

import static camp.nextstep.edu.missionutils.Randoms.pickNumberInRange;

public class Car {
    private String name;
    private int position;

    private Car(String name) {
        this.name = name;
        this.position = 0;
    }

    public static Car create(String name) {
        return new Car(name);
    }

    public static int randomNumberGenerator() {
        int randomNumber = pickNumberInRange(0, 9);
        return randomNumber;
    }

    public void move(int i) {
        if (i >= 4) {
            position += 1;
        }
    }

    public int getPosition() {
        return position;
    }

    public String getName() {
        return name;
    }
}
